debuginfo: extract utils.rs
